Fix bug regarding channel leaving
Scenario: You are in n channels and try to leave the channel which is first on your list **while** having it be the current channel. Background would try to switch to the first channel on your list when trying to leave the current channel, but since both are the same, it would bug out. 

This has been fixed, it will now switch to the next channel beyond that one.